Check auto.cnf file existence (#48)

+++ init-script/run.sh
@@ -5,7 +5,9 @@ rm -rf /run-scripts/*
 cp /tmp/scripts/* /scripts
 if [[ "$PITR_RESTORE" == "true" ]]; then
   if [[ "$HOSTNAME" != *"-0" ]]; then
-    rm /var/lib/mysql/auto.cnf
+    if [[ -f /var/lib/mysql/auto.cnf ]]; then
+          rm /var/lib/mysql/auto.cnf
+    fi
   fi
 fi
